Move pagination logic to dedicated file
